Add files via upload
This app provides quick access to a TESS-W photometer from the STARS4ALL network. It displays the current reading and also a graph of the last 24h by downloading it from the Grafana portal. It has the option to generate an acoustic alarm in case of IR temperature and sky magnitude changes.